refactor: split events.rs into dedicated modules across contracts
